move package handler to /handlers
